requirements.txt
black==21.12b0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
PyMySQL==1.0.2
python-dotenv==0.19.2
tomli==1.2.2
typing-extensions==4.0.1
Werkzeug==2.0.2
